tests: Split readfile tests to run on different runners
These tests each download a 5GB disk image. Running them all in parallel
on one runner was causing download issues. Splitting with one test per runner
should stop the occurance of these issues.
