+++ tests/gem5/readfile_tests/x86/test-readfile.py
@@ -25,48 +25,11 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 """
-This runs simple tests to ensure that running binaries via readfile works.
+This runs simple tests to ensure that running binaries via readfile works for
+X86.
 """
-import os
-import re
+from testlib import constants
 
-from testlib import *
-
-if config.bin_path:
-    resource_path = config.bin_path
-else:
-    resource_path = joinpath(absdirpath(__file__), "..", "resources")
-
-readfile_verifier = verifier.MatchRegex(re.compile(r"Readfile test passed!"))
-
-
-def test_readfile(isa: str, length: str):
-    gem5_verify_config(
-        name=f"test_readfile_{isa}",
-        fixtures=(),
-        verifiers=(readfile_verifier,),
-        config=joinpath(
-            config.base_dir,
-            "tests",
-            "gem5",
-            "readfile_tests",
-            "configs",
-            "ubuntu-run-with-readfile.py",
-        ),
-        config_args=[
-            "--isa",
-            isa,
-            "--resource-directory",
-            resource_path,
-        ],
-        valid_isas=(constants.all_compiled_tag,),
-        valid_hosts=constants.supported_hosts,
-        length=length,
-    )
-
+from gem5.readfile_tests.configs.common import test_readfile
 
 test_readfile(isa="x86", length=constants.long_tag)
-
-test_readfile(isa="riscv", length=constants.long_tag)
-
-test_readfile(isa="arm", length=constants.long_tag)
